Минимальные требования: Приложение принимает твою среднюю зарплату за 12 месяцев и количество дней отпуска - отвечает суммой отпускных, которые придут сотруднику. Доп. задание: При запросе также можно указать точные дни ухода в отпуск, тогда должен проводиться рассчет отпускных с учётом праздников и выходных.

+++ src/main/java/com/example/vacationpaycalculator/controller/CalculateController.java
@@ -22,7 +22,7 @@ public class CalculateController {
             tags = "Калькулятор отпускных"
     )
     @GetMapping("/calculate")
-    public BigDecimal calculate(@RequestParam Double avgSalaryPerMonth,
+    public double calculate(@RequestParam Double avgSalaryPerMonth,
                                 @RequestParam Integer daysOfLeave,
                                 @RequestParam(required = false) List<Integer> daysInLeave) {
         return leaveService.calculateLeavePay(avgSalaryPerMonth, daysOfLeave, daysInLeave);
